+++ Songs-by-Artist/config.py
@@ -1,0 +1,2 @@
+client_id = ""
+secret_key = ""
