--- songs-of-artist/config.py
@@ -1,2 +1,0 @@
-client_id = ""
-secret_key = ""
